requirements.txt
# python==3.8.5
django==3.1.7
pillow==8.2.0
phonenumbers==8.12.21
django-phonenumber-field==5.0.0
babel==2.9.0

